refactor(console): consolidate remote-access pairing and control-handover components
Fold qr-code.tsx into pair-device-dialog.tsx, its only consumer, and merge
control-bar.tsx, control-curtain.tsx and control-reclaimed-notice.tsx into
control-handover.tsx, the concern the existing client test already names.

Mechanical relocation only: every exported symbol is preserved and no private
symbol collided across members.
